requirements.txt
openai>=1.11.0
openai-agents>=0.1.0
fastapi
uvicorn
python-dotenv
duckduckgo-search
